Upgrade to v4.3.0
closes #4
