+++ .devcontainer/docker-compose.yml
@@ -1,5 +1,3 @@
-version: '3.4'
-
 services:
   django:
     env_file:
@@ -36,7 +34,7 @@ services:
       context: ./.devcontainer/docker/geoserver
       dockerfile: Dockerfile
       args:
-        - BASE_IMAGE_VERSION=${GEONODE_VERSION}
+        - BASE_IMAGE_VERSION=${GEOSERVER_BASE_IMAGE_VERSION}
     environment:
       - OAUTH2_CLIENT_ID=${OAUTH2_CLIENT_ID}
       - OAUTH2_CLIENT_SECRET=${OAUTH2_CLIENT_SECRET}
@@ -48,6 +46,6 @@ services:
       context: ./.devcontainer/docker/postgresql
       dockerfile: Dockerfile
       args:
-        - BASE_IMAGE_VERSION=${GEONODE_VERSION}
+        - BASE_IMAGE_VERSION=${POSTGRES_BASE_IMAGE_VERSION}
     ports:
      - "5432:5432"
