+++ .devcontainer/docker/postgresql/Dockerfile
@@ -1,5 +1,5 @@
 ARG BASE_IMAGE_VERSION
-FROM 52north/geonode-postgis:${BASE_IMAGE_VERSION}
+FROM geonode/postgis:${BASE_IMAGE_VERSION}
 
 COPY ./prepare-testdb.sh /docker-entrypoint-initdb.d/prepare-testdb.sh
 RUN chmod +x /docker-entrypoint-initdb.d/prepare-testdb.sh
