+++ docker-compose.yml
@@ -1,14 +1,12 @@
-version: '3.4'
-
 # Common Django template for GeoNode and Celery services below
 x-common-django:
   &default-common-django
-  image: ${DOCKER_REGISTRY}${DOCKER_REPOSITORY}${COMPOSE_PROJECT_NAME}_django:${GEONODE_VERSION}-${PROJECT_VERSION}
+  image: ${DOCKER_REGISTRY}${DOCKER_REPOSITORY}${COMPOSE_PROJECT_NAME}_django:${GEONODE_BASE_IMAGE_VERSION}-${PROJECT_VERSION}
   build:
     context: ./docker/geonode
     dockerfile: Dockerfile
     args:
-      - BASE_IMAGE_VERSION=${GEONODE_VERSION}
+      - BASE_IMAGE_VERSION=${GEONODE_BASE_IMAGE_VERSION}
   restart: no
   env_file:
     - .env
@@ -71,7 +69,7 @@ services:
       context: ./docker/nginx
       dockerfile: Dockerfile
       args:
-        - BASE_IMAGE_VERSION=${GEONODE_VERSION}
+        - BASE_IMAGE_VERSION=${NGINX_BASE_IMAGE_VERSION}
     container_name: nginx4${COMPOSE_PROJECT_NAME}
     env_file: ./.env
     environment:
@@ -99,7 +97,7 @@ services:
       context: ./docker/geoserver
       dockerfile: Dockerfile
       args:
-        - BASE_IMAGE_VERSION=${GEONODE_VERSION}
+        - BASE_IMAGE_VERSION=${GEOSERVER_BASE_IMAGE_VERSION}
     container_name: geoserver4${COMPOSE_PROJECT_NAME}
     env_file: ./.env
     healthcheck:
@@ -128,7 +126,7 @@ services:
       context: ./docker/geoserver_data
       dockerfile: Dockerfile
       args:
-        - BASE_IMAGE_VERSION=${GEONODE_VERSION}
+        - BASE_IMAGE_VERSION=${GEOSERVER_DATA_BASE_IMAGE_VERSION}
     container_name: gsconf4${COMPOSE_PROJECT_NAME}
     entrypoint: sleep infinity
     volumes:
@@ -145,7 +143,7 @@ services:
       context: ./docker/postgresql
       dockerfile: Dockerfile
       args:
-        - BASE_IMAGE_VERSION=${GEONODE_VERSION}
+        - BASE_IMAGE_VERSION=${POSTGRES_BASE_IMAGE_VERSION}
     command: postgres -c "max_connections=${POSTGRES_MAX_CONNECTIONS}"
     container_name: db4${COMPOSE_PROJECT_NAME}
     env_file: ./.env
