+++ docker/geoserver/Dockerfile
@@ -1,5 +1,5 @@
 ARG BASE_IMAGE_VERSION
-FROM 52north/geonode-geoserver:${BASE_IMAGE_VERSION}
+FROM geonode/geoserver:${BASE_IMAGE_VERSION}
 
 ## GeoServer plugins
 #
