+++ docker/nginx/Dockerfile
@@ -1,5 +1,5 @@
 ARG BASE_IMAGE_VERSION
-FROM 52north/geonode-nginx:${BASE_IMAGE_VERSION}
+FROM geonode/nginx:${BASE_IMAGE_VERSION}
 
 COPY ./cert.pem /geonode-certificates/autoissued/fullchain.pem
 COPY ./key.pem /geonode-certificates/autoissued/privkey.pem
