+++ geonode/geonode/settings_override.py
@@ -8,13 +8,18 @@ import os
 import sys
 
 
-# DEBUG = "False"
-
-X_FRAME_OPTIONS = "SAMEORIGIN"
-
 # sets defaults settings and from .env
 from geonode.settings import *
-from geonode.settings import TEMPLATES, INSTALLED_APPS
+from geonode.settings import (
+    DEBUG,
+    TEMPLATES,
+    INSTALLED_APPS,
+)
+
+X_FRAME_OPTIONS = "SAMEORIGIN"
+SECURE_CROSS_ORIGIN_OPENER_POLICY = None if DEBUG else "same-origin"
+# required for geonode-mapstore-client development
+CSRF_TRUSTED_ORIGINS = ["http://localhost:8081"] if DEBUG else []
 
 
 STATIC_ROOT = "/mnt/volumes/statics/static/"
@@ -105,6 +110,6 @@ LOGGING = {
 
 
 INSTALLED_APPS += (
-    "externalapplications",
+    # "externalapplications",
     "customizations",
 )
