+++ sample.env
@@ -4,12 +4,18 @@ COMPOSE_PROJECT_NAME="geonode_docker"
 DOCKER_REGISTRY=
 # (optional) with trailing slash e.g. sample/ or orga/
 DOCKER_REPOSITORY=my-geonode/
-GEONODE_VERSION="4.2.3"
+GEONODE_BASE_IMAGE_VERSION="4.3.0"
+NGINX_BASE_IMAGE_VERSION="1.25.3-v1"
+GEOSERVER_BASE_IMAGE_VERSION="2.24.3-v1"
+GEOSERVER_DATA_BASE_IMAGE_VERSION="2.24.3-v1"
+POSTGRES_BASE_IMAGE_VERSION="15.3-v1"
+
 PROJECT_VERSION="v1"
 
 
 # set to 'true' to re-initialize setup
 FORCE_REINIT=false
+INVOKE_LOG_STDOUT=True
 
 
 ## External Site Settings
